+++ react/country/ARG.js
@@ -21310,6 +21310,10 @@ export default {
     [{ name: 'street' }, { delimiter: ' ', name: 'number' }],
     [{ name: 'complement' }],
     [{ name: 'postalCode' }],
-    [{ name: 'city' }, { delimiter: ', ', name: 'state' }],
+    [
+      { name: 'neighborhood', delimiterAfter: ' - ' },
+      { name: 'city' },
+      { delimiter: ', ', name: 'state' },
+    ],
   ],
 }